Modify IView interface. From update(Model m, Event e) to update(Event e) in order to reduce workload on the net. It avoid sending always model to the view. The view it's responsible of updating the model if necessary.
Change PatternCardDistributeEvent in order to carry the PatternCards that has to be chosen.

Add IToolCardFiller interface. The class's concerns are related to the proper creation of toolcards. A object of a class that inherits from Toolcard is responsible of transporting the necessary information to get it applied to the model.The cards are filled from the view, then are passed from the controller to the model, which tries to apply it, calling applyEffect giving a model and a player reference.

Add fill as an entrypoint of the visitor pattern represented by IToolCardFiller

Review GameManager
Review Player

Add factory method to constraint