Offline game [fixed]
- The max nr of player have to be chosen during the creation of the controller by passing a int to its constructor. (in the future a timeout should be provided)

MATCH CREATION FLOW
The controller contains the game obj. A reference (remote reference with RMI) to the controller must be provided to each view (client); then the view can join the game calling joinTheGame(myName, myView) on the controller obj. As soon as the internal game in the controller reaches the desired nr of players (2, 3 or 4) it automatically creates a gamemanager (with the GameManagerFactory object) and keeps a reference to it. Then the controller invokes the setupPhase() on the newly created GameManerger. From now on the match starts normally :)