Add dice graphics
modify player / game in order to not disturb ObjectOutput/InputStream. This was done primary because of the not known behaviour of the ObjectOutputStream. Professor said that when we try to serialize with objectOutputStream that has already been serialized the systems recognize and instead of serializing the whole object it and uses a simpler encoding, not wasting band on the network. The problem is that we actually don't the stream recognizes the objects (using equals in hashmap / using references). So i override the hashCode, equals of game/gamamanager in order to not considering the reference anymore.

Fix a small problem in controller bindPattern which checked for currentPlayer -> if exist a player that uses that name the operation is considered legitimate.

Controller was made synchronized (i.e. every method)

Introduce response/request interfaces/classes to accomplish a sort of remote method calling realizing 2 classes that act as a proxy for both controller (ViewProxy) and the view (ControllerProxy).
Since the communication between server/client is deeply asynchronous the endpoints of the communications can both receive and send messages without having to rely on a protocol which states that after a request comes always a response (like in @affo chat implementation).

Note that a new view was introduced in order to handle asynchronism problems. ProxyController and the view (i.e. ClientViewCLI) transmit data from/to the server invoking respectively update(Event aEvent), handleException(Exception aException) and the methods indicated into IControllerInterface. Every operation, i.e. invocation of method, is translated into a push of an event/request into a queue and a signal to the thread responsible for the handling. This is primary because some resources have to be managed from only a thread from its birth to its death. Because of this the thread managing these resources can't block themselves during operation. This was accomplished using Socket::setSoTimeout() and BuffredReader/preemptiveReadline/InterruptException respectively in the Proxy and in the ClientViewCLI.

Hope it helps